pyproject.toml
[project]
name = "viral-clip-generator-system"
version = "0.1.0"
description = "AI-powered video processing tool that intelligently combines clips to create unlimited viral content"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0"
]
